feat(epl): port EPLVariablesUse final 2 runtimes. Coverage 554/4136 cases (167 DV, 652 DV runtime IDs)
case.variables-use closes EPLVariablesUse at 9/11 executions (101/101
records, 0 differences): EPVariableService runtime API surface and
constant-variable truth tables. Engine: Java-widening coercion chain,
verbatim runtime set-path diagnostics, per-slot multi-match IN delivery
for collection candidates (Java FilterParamIndexIn flattened slots),
map-payload struct materialization with string->named-kind coercion,
membership matching via equalValuesUnwrapped. Oracle/runner: persistent
compiled-module path with SODA re-create bypass, targeted undeploy op,
label-keyed deployments. Registered unrepresented: DotSeparateThread,
WVarargs, byte[] boxed/primitive distinction, SupportBean[] array
rejection. Review PASS.

+++ internal/esper/database_join_parity_test.go
@@ -877,8 +877,8 @@ func TestDatabaseNoJoinVariablesPollMatchesJava(t *testing.T) {
 
 	if err := engine.SetVariables(context.Background(),
 		VariableAssignment{Name: "queryvar_bool", Value: true},
-		VariableAssignment{Name: "lower", Value: int64(10)},
-		VariableAssignment{Name: "upper", Value: int64(40)},
+		VariableAssignment{Name: "lower", Value: 10},
+		VariableAssignment{Name: "upper", Value: 40},
 	); err != nil {
 		t.Fatal(err)
 	}
@@ -904,8 +904,8 @@ func TestDatabaseNoJoinVariablesPollMatchesJava(t *testing.T) {
 
 	if err := engine.SetVariables(context.Background(),
 		VariableAssignment{Name: "queryvar_bool", Value: false},
-		VariableAssignment{Name: "lower", Value: int64(30)},
-		VariableAssignment{Name: "upper", Value: int64(80)},
+		VariableAssignment{Name: "lower", Value: 30},
+		VariableAssignment{Name: "upper", Value: 80},
 	); err != nil {
 		t.Fatal(err)
 	}
